+++ propulsion-summary-consumer/TodoSummary.fs
@@ -18,23 +18,8 @@ module Folds =
     let evolve _state = function
         | Events.Ingested e -> { version = e.version; value = Some e.value }
     let fold : State -> Events.Event seq -> State = Seq.fold evolve
-    let private isOrigin = function Events.Ingested _ -> true
-    // A `transmute` function gets presented with:XX
-    // a) events a command decided to generate (in it's `interpret`)
-    // b) the state after applying them
-    // and is expected to return:
-    // a) the revised list of events to actually write as events in the stream
-    // b) the snapshot(s) to put in the `u`nfolds list in the Tip
-    //
-    // This implementation means that every time `interpret` decides to write an `Ingested` event, we flip what would
-    // normally happen (write a new event in a new document in the stream and update the snapshot so we can read it in one shot)
-    // and use AccessStrategy.RollingUnfolds with this `transmute` function so we instead convey:
-    // a) "don't actually write these events we just decided on in `interpret` [and don't insert a new event batch document]"
-    // b) "can you treat these events as snapshots please"
-    let private transmute events _state : Events.Event list * Events.Event list =
-        [],events
-    // We don't want to write any events, so here we supply the `transmute` function to teach it how to treat our events as snapshots
-    let accessStrategy = Equinox.Cosmos.AccessStrategy.RollingUnfolds (isOrigin,transmute)
+    let snapshot state = Events.Ingested { version = state.version; value = state.value.Value }
+    let accessStrategy = Equinox.Cosmos.AccessStrategy.RollingState snapshot
 
 module Commands =
     type Command =